fastboot: Skip sparse image conversion if size is within limit
As it turns out, the bootloader will actually accept raw images, not
just sparse ones. Sparse images are only needed for splitting raw images
that are too big for a single download+flash transaction. In fact,
flashing avb_custom_key breaks when we convert it to sparse because the
bootloader expects raw RSA public key data (with an AVB header) in it.

This also removes the BOOTLDR! magic check because it's no longer
necessary with the new logic.