demo: download: Remove dependency on lib/common.js
The dependency on lib/common.js is a relic from when BlobStore was part
of the fastboot.js library. We can't import it directly anymore now that
it has been migrated to TypeScript. It doesn't make much sense to use
the library's internal logging facility here anyway, so remove the
dependency and use console.log directly.